Use project_config.name as default sec_group_name
This hacks in a 'base_name' param to the config dict that
represents the raw 'name' provided in the config. This base_name
is used as the sec_group_name in the case that no specific
sec_group_name is provided. This allows us to better separate
security group rules by project. This purposefully doesn't use
the env name passed through the CLI as a part of the sec_group_name.

Fixes issue #62